feature: hotwire combobox
