+++ app/javascript/js/application.js
@@ -1,5 +1,6 @@
 import { Alert, Popover } from 'tailwindcss-stimulus-components'
 import { Application } from '@hotwired/stimulus'
+import HwComboboxController from '@josefarias/hotwire_combobox'
 import TurboPower from 'turbo_power'
 
 TurboPower.initialize(window.Turbo.StreamActions)
@@ -13,5 +14,6 @@ window.Stimulus = application
 // Register stimulus-components controller
 application.register('alert', Alert)
 application.register('popover', Popover)
+application.register('hw-combobox', HwComboboxController)
 
 export { application }
