+++ app/views/layouts/avo/application.html.erb
@@ -22,6 +22,7 @@
         <%= javascript_include_tag "hotwire-livereload", defer: true %>
       <% end %>
     <% end %>
+    <%= combobox_style_tag %>
     <%= render Avo::AssetManager::JavascriptComponent.new asset_manager: Avo.asset_manager %>
     <%= render partial: 'avo/partials/head' %>
     <%= turbo_refreshes_with method: :replace, scroll: :reset %>
