+++ package.json
@@ -24,6 +24,7 @@
     "@babel/plugin-proposal-class-properties": "^7.12.1",
     "@hotwired/stimulus": "^3.2.2",
     "@hotwired/turbo-rails": "^8.0.4",
+    "@josefarias/hotwire_combobox": "^0.2.2",
     "@rails/activestorage": "^6.1.7",
     "@tailwindcss/forms": "^0.5.7",
     "@tailwindcss/typography": "^0.5.12",
